+++ src/parser/inline/autolink/account.test.ts
@@ -20,11 +20,11 @@ describe('Unit: parser/inline/autolink/account', () => {
       assert.deepStrictEqual(inspect(parser('@@')), [['@@'], '']);
       assert.deepStrictEqual(inspect(parser('@@a')), [['@@'], 'a']);
       assert.deepStrictEqual(inspect(parser('a@b')), [['a@'], 'b']);
-      assert.deepStrictEqual(inspect(parser('a@http://host')), [['a@ht'], 'tp://host']);
-      assert.deepStrictEqual(inspect(parser('a@ttp://host')), [['a@t'], 'tp://host']);
+      assert.deepStrictEqual(inspect(parser('a@http://host')), [['a@'], 'http://host']);
+      assert.deepStrictEqual(inspect(parser('a@ttp://host')), [['a@'], 'ttp://host']);
       assert.deepStrictEqual(inspect(parser('a@@b')), [['a@@'], 'b']);
-      assert.deepStrictEqual(inspect(parser('a@@http://host')), [['a@@ht'], 'tp://host']);
-      assert.deepStrictEqual(inspect(parser('a@@ttp://host')), [['a@@t'], 'tp://host']);
+      assert.deepStrictEqual(inspect(parser('a@@http://host')), [['a@@'], 'http://host']);
+      assert.deepStrictEqual(inspect(parser('a@@ttp://host')), [['a@@'], 'ttp://host']);
     });
 
     it('valid', () => {
@@ -65,6 +65,7 @@ describe('Unit: parser/inline/autolink/account', () => {
       assert.deepStrictEqual(inspect(parser('@a_')), [['<span class="account">@a</span>'], '_']);
       assert.deepStrictEqual(inspect(parser('@a$')), [['<span class="account">@a</span>'], '$']);
       assert.deepStrictEqual(inspect(parser('@a@')), undefined);
+      assert.deepStrictEqual(inspect(parser('@a#')), [['<span class="account">@a</span>'], '#']);
       assert.deepStrictEqual(inspect(parser('@a[')), [['<span class="account">@a</span>'], '[']);
       assert.deepStrictEqual(inspect(parser('@a]')), [['<span class="account">@a</span>'], ']']);
       assert.deepStrictEqual(inspect(parser('@a(')), [['<span class="account">@a</span>'], '(']);