+++ src/parser/inline/autolink/account.ts
@@ -1,13 +1,11 @@
 ﻿import { AutolinkParser } from '../../inline';
 
 const syntax = /^@[a-zA-Z0-9]+(?:-[0-9a-zA-Z]+)*(?!@)/;
-const escape = /^[0-9a-zA-Z@]@/;
+const escape = /^[0-9a-zA-Z@]@.*?(?!@)/;
 
-export const account: AutolinkParser.AccountParser = (source: string): [[HTMLSpanElement | Text], string] | undefined => {
-  if (source.search(escape) === 0) {
-    const [frag = source] = source.match(/^[0-9a-zA-Z@].*?(?!@|h?ttps?:)/) || [];
-    return [[document.createTextNode(frag)], source.slice(frag.length)];
-  }
+export const account: AutolinkParser.AccountParser = (source: string) => {
+  const [flag = undefined] = source.match(escape) || [];
+  if (flag) return [[document.createTextNode(flag)], source.slice(flag.length)];
   const [whole = ''] = source.match(syntax) || [];
   if (!whole) return;
   const el = document.createElement('span');